+++ structs/response.go
@@ -4,7 +4,7 @@ import "time"
 
 // Response is a struct that stores information about a http request response.
 type Response struct {
-	Duration   int64
+	Duration   int
 	Error      bool
 	Size       int64
 	Start      time.Time
